:recycle: refactor bggSearch related features
