+++ src/features/bgg-search/components/bgg-search/bgg-search.component.scss
@@ -1,4 +1,4 @@
-@use "../../core/styles/variables" as variables;
+@use "../../../../core/styles/variables" as variables;
 
 p {
   margin: 2px;
